feat(promql): add latestk(k, series{}) aggregate
The latestk aggregate is used to select the series with the most
recent sample timestamps within each aggregation group.

It is expected to be mostly usedful for k=1 for the purpose of finding
the series with the most recent sample within a group of series that are
used to represent an enumerated set (StatefulSet).

Traditionally, these are represented with 0-valued series for "not the
active state" to avoid various issues with staleness markers being
missed.

By using

    latestk(1, series{} == 1) by (identifing_labels)

it becomes possible reliably determine the most recent state for series
that omit emission/ingestion of those 0-valued samples to represent
inactive states. Including the == 1 filter ensures that the pattern also
works for series that do still include 0-valued inactive state series.