promql: Add new timestamp_if_value(vec, val) function
The timestamp_if_value(vec vector, val scalar) function returns the
timestamps of the input series in 'vec' only if the sample value matches
scalar literal 'val'.

Evaluating expressions to expand 'val' is not currently supported, and
will raise an error at runtime.

This is a proof of concept for #17129 and is subject to change in
design, name, etc.

The primary purpose of this function is to provide a means to reliably
handle samples from enumerated sets when the series cannot be relied
upon to have 0-valued samples or stale markers for inactive states.

See the added documentation and test cases for a detailed explanation
and examples.